Product correction. replaced Double to BigDecimal in Product entity.

+++ src/main/java/com/homecomfort/homecomfort/entity/Product.java
@@ -2,6 +2,8 @@ package com.homecomfort.homecomfort.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "products")
 public class Product {
@@ -14,7 +16,7 @@ public class Product {
     private String name;
 
     @Column(nullable = false)
-    private Double price;
+    private BigDecimal price;
 
     @Column(length = 1000)
     private String description;
@@ -26,7 +28,7 @@ public class Product {
     public Product() {
     }
 
-    public Product(Long id, String name, Double price, String description, Category category) {
+    public Product(Long id, String name, BigDecimal price, String description, Category category) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -42,7 +44,7 @@ public class Product {
         return name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
@@ -62,7 +64,7 @@ public class Product {
         this.name = name;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
